neural_engine/requirements.txt
# Nanobot Engine - Core Dependencies
# ====================================

# API Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-multipart==0.0.19
pydantic==2.10.4
pydantic-settings==2.7.1

# LLM Integration — Hybrid Architecture
ollama==0.4.7
httpx==0.28.1
groq>=1.0.0               # Groq cloud API (llama-3.3-70b-versatile)
litellm>=1.81.0            # Multi-provider LLM gateway
nanobot-ai>=0.1.4          # Official nanobot framework (channel gateway)

# Browser Automation (Persistent Sessions)
playwright==1.49.1

# MCP (Model Context Protocol)
mcp==1.3.0

# Calendar & Email
icalendar==6.1.1
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Knowledge / Notes
markdown==3.7
python-frontmatter==1.1.0

# Utilities
python-dotenv==1.0.1
rich==13.9.4
watchdog==6.0.0
aiofiles==24.1.0
orjson==3.10.14
